requirements.txt
pylast~=5.0.0
python-dotenv~=0.20.0
rymscraper~=0.2.3
